+++ BambooTracker/command/pattern/set_effect_value_to_step_command.cpp
@@ -28,6 +28,7 @@ void SetEffectValueToStepCommand::undo()
 {
 	mod_.lock()->getSong(song_).getTrack(track_).getPatternFromOrderNumber(order_)
 			.getStep(step_).setEffectValue(n_, prevVal_);
+	isSecond_ = true;	// Forced complete
 }
 
 CommandId SetEffectValueToStepCommand::getID() const