+++ BambooTracker/gui/command/pattern/set_effect_id_to_step_qt_command.cpp
@@ -17,6 +17,7 @@ void SetEffectIDToStepQtCommand::redo()
 void SetEffectIDToStepQtCommand::undo()
 {
 	panel_->redrawByPatternChanged();
+	panel_->resetEntryCount();
 }
 
 int SetEffectIDToStepQtCommand::id() const