+++ BambooTracker/gui/command/pattern/set_effect_value_to_step_qt_command.cpp
@@ -17,6 +17,7 @@ void SetEffectValueToStepQtCommand::redo()
 void SetEffectValueToStepQtCommand::undo()
 {
 	panel_->redrawByPatternChanged();
+	panel_->resetEntryCount();
 }
 
 int SetEffectValueToStepQtCommand::id() const