+++ BambooTracker/gui/order_list_editor/order_list_panel.hpp
@@ -46,6 +46,8 @@ public:
 	void redrawBySizeChanged();
 	void redrawAll();
 
+	void resetEntryCount();
+
 	void freeze();
 	void unfreeze();