+++ BambooTracker/gui/pattern_editor/pattern_editor_panel.hpp
@@ -51,6 +51,8 @@ public:
 	void redrawByHeaderChanged();
 	void redrawBySizeChanged();
 
+	void resetEntryCount();
+
 	void freeze();
 	void unfreeze();