+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Disable slider operation during playback
 - Fix corruption after `0Bxx`, `0Cxx` and `0Dxx` are set the position before current playback row
 - Fix incorrect square-mask pitch in square-masked saw and inversed saw
+- Reset character entry position in pattern editor and order list after undoing
 
 [#152]: https://github.com/rerrahkr/BambooTracker/pull/152
 [#153]: https://github.com/rerrahkr/BambooTracker/issues/153